backend/requirements.txt
# Core dependencies
fastapi>=0.103.0
uvicorn>=0.23.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Azure services
azure-storage-blob>=12.17.0
azure-cosmos>=4.5.0

# Data handling
python-multipart>=0.0.6  # For form data handling
types-python-dateutil>=2.8.0  # Type stubs

# Security & validation
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0  # JWT support (if needed)
passlib[bcrypt]>=1.7.4  # Password hashing (optional)

# Optional performance
uvloop>=0.17.0  # Faster ASGI server (Linux only)
httptools>=0.6.0  # HTTP parser
orjson>=3.8.0  # Faster JSON parsing


# For development/testing only
pytest>=7.0.0
requests>=2.31.0
pytest-asyncio>=0.21.0
httpx>=0.23.0
freezegun>=1.2.0
pandas-stubs>=1.2.0  # Type hints for pandas